move member functions to their own file first
